requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
anthropic==0.32.0
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
pyyaml==6.0.1
fake-useragent==1.5.1
